# swagger.yaml
openapi: 3.0.0
info:
  title: SIT POS Microservices API
  description: API Gateway for Authentication and Merchant Services
  version: 1.0.0
  contact:
    name: API Support
    email: support@sitpos.com
servers:

  - url: https://unemotionally-unwayward-brandie.ngrok-free.dev
    description: API Gateway (NGROK)
  - url: http://192.168.2.106:3000
    description: API Gateway (Local Network)
  - url: http://localhost:3000
    description: API Gateway (Localhost)
  - url: http://localhost:3001
    description: Auth Service (Development)
  - url: http://localhost:3002
    description: Merchant Service (Development)
  - url: http://localhost:3003
    description: Catalog Service (Development)
  - url: http://localhost:3004
    description: Cart Service (Development)
...
components:
  securitySchemes:
    ServiceToken:
      type: apiKey
      in: header
      name: X-Service-Token
      description: |
          Service-to-service authentication token.
 
          **For Swagger Testing, use this token:**
          ```eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzZXJ2aWNlIjoic3dhZ2dlci10ZXN0IiwidHlwZSI6InNlcnZpY2UtdG8tc2VydmljZSIsImlhdCI6MTc1OTk4NjkyNywiZXhwIjoxNzYwMDczMzI3fQ.4PwIvt746-tRzaT3-EoItg5kvAElv73xwLXjnbL1YvU```
 
          Paste this token below and click on "Authorize" button below.

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    Merchant:
      type: object
      required:
        - merchantId
        - businessInfo
      properties:
        merchantId:
          type: string
          example: "MERCH_001"
        businessInfo:
          type: object
          required:
            - legalName
          properties:
            legalName:
              type: string
              example: "Test Coffee Shop LLC"
            dbaName:
              type: string
              example: "Joe's Coffee"
            businessType:
              type: string
              example: "LLC"
            industry:
              type: string
              example: "Food & Beverage"
            taxId:
              type: string
              example: "12-3456789"
        address:
          type: object
          properties:
            street:
              type: string
              example: "123 Main Street"
            city:
              type: string
              example: "San Francisco"
            state:
              type: string
              example: "CA"
            zipCode:
              type: string
              example: "94102"
            country:
              type: string
              example: "US"
              default: "US"
        contact:
          type: object
          properties:
            primaryPhone:
              type: string
              example: "+14155551234"
            primaryEmail:
              type: string
              format: email
              example: "contact@joescoffee.com"
        owner:
          type: object
          properties:
            firstName:
              type: string
              example: "Joe"
            lastName:
              type: string
              example: "Smith"
            email:
              type: string
              format: email
              example: "joe@joescoffee.com"
            phone:
              type: string
              example: "+14155551234"
        status:
          type: string
          enum: [pending, active, inactive, suspended]
          default: pending
          example: "active"
...
paths:
...
  /api/v1/merchant:
    post:
      tags:
        - Merchants
      summary: Create merchant
      description: Create a new merchant (requires service token)
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Merchant'
      responses:
        '201':
          description: Merchant created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Merchant'

    get:
      tags:
        - Merchants
      summary: List merchants
      description: Get list of all merchants (requires service token)
      security:
        - ServiceToken: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, active, inactive, suspended]
        - in: query
          name: limit
          schema:
            type: integer
            example: 10
        - in: query
          name: offset
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: List of merchants
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Merchant'
...
  /api/v1/device:
    get:
      tags:
        - Devices
      summary: List devices
      description: Get list of all devices (requires service token)
      security:
        - ServiceToken: []
      responses:
        '200':
          description: List of devices
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
...
  /api/v1/cart:
    post:
      tags:
        - Cart
      summary: Get or create cart
      description: Get or create active cart for customer
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - merchant_id
                - customer_id
              properties:
                merchant_id:
                  type: string
                  example: "MERCHANT_1234567890"
                customer_id:
                  type: string
                  example: "CUSTOMER_987654321"
                device_id:
                  type: string
                  example: "DEVICE_ABC123XYZ"
      responses:
        '200':
          description: Cart retrieved or created successfully
...